% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidplyrio-package.R
\docType{package}
\name{multidplyrio-package}
\alias{multidplyrio}
\alias{multidplyrio-package}
\title{multidplyrio: Extend Multidplyr with I/O functions}
\description{
Provide additional I/O functions for use with the {multidplyr} package, to aid reading data into workers quickly.
}
\author{
\strong{Maintainer}: Isaac Florence \email{isaac.florence@outlook.com} (\href{https://orcid.org/0000-0001-8260-1237}{ORCID})

}
\keyword{internal}
