% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-delim.R
\name{read_delim_distrib}
\alias{read_delim_distrib}
\title{Read delimited files to workers}
\usage{
read_delim_distrib(
  cluster,
  filename,
  delim = ",",
  col_types = NULL,
  show_col_types = TRUE
)
}
\arguments{
\item{cluster}{a multidplyr cluster}

\item{filename}{a path and file of the delimited file to read onto workers}

\item{delim}{Single character used to separate fields within a record.}

\item{col_types}{\code{NULL} or a \code{\link[=cols]{cols()}} specification.
See \code{vignette("readr")} for more details. Note that current strings with
compact col_type specifications are not supported.

If \code{NULL}, all column types will be imputed from 200 rows
on the input interspersed throughout the file. This is convenient (and
fast), but not robust. If the imputation fails, you'll need to supply the
correct types yourself.

Column specifications created by \code{\link[=list]{list()}} or \code{\link[=cols]{cols()}} must contain
one column specification for each column. If you only want to read a
subset of the columns, use \code{\link[=cols_only]{cols_only()}}.}

\item{show_col_types}{Set to \code{FALSE} to silence col_type summary printing}
}
\description{
Read a delimited file (e.g. a csv) onto \strong{multidplyr} workers in parallel
powered by \strong{readr} and \strong{vroom}. For most arguments below, the parameters
are applied directly to - and documentation adapted from - readr::read_delim
}
